+++ src/etcdExplorer.ts
@@ -12,16 +12,28 @@ export class EtcdExplorerBase {
   protected client: any;
   protected conflictsResolution = "abort";
   protected sameKeysResolution = "abort";
+  protected arrayResolution = "Array as string";
+  private context: vscode.ExtensionContext;
   protected documentChanged?: vscode.EventEmitter<vscode.Uri>;
-  constructor(schema: string) {
+  constructor(schema: string, _context: vscode.ExtensionContext) {
     console.log("Constructing ETCD Explorer");
+    this.context = _context;
     this.etcdSch = schema;
+    this.reloadConfig();
+    this.rootNode = new EtcdRootNode(this);
+  }
+
+  protected reloadConfig() {
     var conf = vscode.workspace.getConfiguration('etcd-explorer');
     var importJSON = conf.importJSON;
-    console.log(importJSON);
     this.conflictsResolution = importJSON.conflicts;
     this.sameKeysResolution = importJSON.sameKeys;
-    this.rootNode = new EtcdRootNode(this);
+    this.arrayResolution = importJSON.arrays;
+  }
+
+  protected getConfig() {
+    var conf = vscode.workspace.getConfiguration('etcd-explorer');
+    return conf;
   }
 
   setDocumentChangedEventEmitter(_documentChanged?: vscode.EventEmitter<vscode.Uri>) {
@@ -46,6 +58,12 @@ export class EtcdExplorerBase {
   }
 
   refreshView(clusterHost: string) {
+    var conf = vscode.workspace.getConfiguration('etcd-explorer');
+    var importJSON = conf.importJSON;
+    this.conflictsResolution = importJSON.conflicts;
+    this.sameKeysResolution = importJSON.sameKeys;
+    this.arrayResolution = importJSON.arrays;
+
     if (clusterHost != this.etcd_host) {
       this.etcd_host = clusterHost;
       this.refreshData();
@@ -54,79 +72,79 @@ export class EtcdExplorerBase {
 
   getChildren(element?: EtcdNode): Thenable<EtcdNode[]> {
     element = element ? element : this.rootNode;
-    var children = element.getChildren(true);
-    children.removeUpdatingNodes();
-    if (children.updatingNodes) {
+    var children = element.getChildren();
+    if (children.isUpdating()) {
       var arrayChildren = children.toArray();
-      arrayChildren.push(new EtcdUpdatingNode(this, this.rootNode));
+      if (children.getNodeByLabel("loading").leaf === undefined)
+        arrayChildren.push(new EtcdUpdatingNode(this, element));
       return Promise.resolve(arrayChildren);
     }
+    children.removeUpdatingNodes();
     if (this.rootNode.getChildren(true).toArray().length == 0) {
       return Promise.resolve([new EtcdEmptyWSNode(this, this.rootNode)]);
     }
     return Promise.resolve(element.getChildren().toArray());
   }
 
-  protected write(key: string, value: any) { }
+  protected async write(key: string, value: any) { return new Promise((resolve, reject) => { reject("Not Implemented in base class"); }); }
 
-  async jsonToEtcd(jsonObj: JSON) {
+  // used by importJSON and addKeyValue
+  async jsonToEtcd(jsonObj: JSON, token?: vscode.CancellationToken) {
     if (this.client === undefined) return;
     var currentKey = "";
     var writes = new Array<{ value: string, key: string }>();
     var conflicts = new Array<{ value: string, key: string, srcLeaf: boolean, dstLeaf: boolean }>();
+    this.reloadConfig();
+    // using stack to recurse through JSON object
     var stack = new Array<{ json: JSON, key: string }>();
     stack.push({ json: jsonObj, key: currentKey });
     while (stack.length > 0) {
+      if (token && token.isCancellationRequested) return;
       var obj = stack.pop();
       if (!obj) break;
       var json = obj.json;
       var entries = Object.entries(json);
       for (var entry of entries) {
+        if (token && token.isCancellationRequested) return;
         currentKey = obj.key;
         var isLeaf = false;
         var key = entry[0];
         var value = entry[1];
-        var str = Object.prototype.toString.call(value);
-        var type = str.slice(8, -1).toLowerCase();
-        if (type == "string"
-          || type == "boolean"
-          || type == "number"
-          || type == "date"
-          || type == "array"
-          || type == "undefined"
-          || type == "null"
-        ) {
-          isLeaf = true;
-        }
-        else {
-          isLeaf = false;
-        }
-        currentKey += separator + key;
+        isLeaf = this.isValueType(value);
+        currentKey += key.startsWith(separator) ? key : (separator + key);
         var existingNode = this.findEtcdNode(currentKey);
         if (!existingNode) {
           existingNode = this.findEtcdNode(currentKey + separator);
         }
-        if (!isLeaf) {
-          if (existingNode && existingNode.isLeafNode()) {
-            conflicts.push({ key: currentKey, value: value, srcLeaf: isLeaf, dstLeaf: existingNode.isLeafNode() });
+        if (!isLeaf) {  /*json key is not leaf but dir*/
+          if (existingNode && existingNode.isLeafNode()) { /* existing node is not dir but leaf*/
+            // this is allowed in etcd3 but not in etcd2 
+            // etcd3 can have cases like: /a/b = c && /a = e
+            if (this.schema().startsWith('etcd2')) {
+              conflicts.push({ key: currentKey, value: value, srcLeaf: isLeaf, dstLeaf: existingNode.isLeafNode() });
+            }
           }
           stack.push({ json: value, key: currentKey });
         }
-        else {
+        else { /*json key is leaf */
           if (existingNode) {
-            if (!existingNode.isLeafNode()) {
-              conflicts.push({ key: currentKey, value: value, srcLeaf: isLeaf, dstLeaf: existingNode.isLeafNode() });
-            }
             conflicts.push({ key: currentKey, value: value, srcLeaf: isLeaf, dstLeaf: existingNode.isLeafNode() });
           }
           else {
-            if (type == "array") {
-              var index = 0;
-              for (var val of value) {
-                var valObj = Object.create({});
-                valObj["[" + index + "]"] = val;
-                index++;
-                stack.push({ key: currentKey, json: valObj });
+            if (this.getType(value) == "array") {
+              this.reloadConfig();
+              if (this.arrayResolution == "Array as string") {
+                value = JSON.stringify(value);
+                writes.push({ key: currentKey, value: value });
+              }
+              else {
+                var index = 0;
+                for (var val of value) {
+                  var valObj = Object.create({});
+                  valObj["[" + index + "]"] = val;
+                  index++;
+                  stack.push({ key: currentKey, json: valObj });
+                }
               }
             }
             else {
@@ -137,27 +155,64 @@ export class EtcdExplorerBase {
       }
     }
     if (conflicts.length > 0) {
+      var conflictStr = "";
+      var abortImport = false;
       for (var conflict of conflicts) {
-        if (conflict.srcLeaf != conflict.dstLeaf) {
-          if (this.conflictsResolution == "abort") {
-            vscode.window.showErrorMessage("Aborting json import due to conflict with existing keys");
-            var p = new Promise(async (resolve) => {
-              var str = "Conflicts:\n ==================================\n";
-              for (var c of conflicts) {
-                str += c.key + ": " + c.value + "\n";
+        conflictStr = "Conflicts:\n ==================================\n";
+        if (conflict.srcLeaf && !conflict.dstLeaf) {
+          conflictStr += ">>>>Existing key is not a folder\n" + conflict.key + ": " + conflict.value + "\n";
+          if (this.conflictsResolution == "abort") abortImport = true;
+          if (this.conflictsResolution == "overwrite") {
+            var response: string | undefined;
+            var noOwPromptCR = this.context.globalState.get("no_overwrite_prompt_for_conflicting_keys");
+            if (!noOwPromptCR) {
+              await vscode.window.showWarningMessage("Etcd Explorer: You have chosen to overwrite the values for same keys in settings, do you wish to continue?", { "modal": true }, "Yes, and don't ask again", "Yes").then((s) => { response = s; });
+              if (response === undefined) abortImport = true;
+              else if (response != "Yes") {
+                this.context.globalState.update("no_overwrite_prompt_for_conflicting_keys", true);
               }
-              let doc = await vscode.workspace.openTextDocument({ content: str, language: "json" });
-              vscode.window.showTextDocument(doc, { preview: false });
-              resolve();
-            });
-
-            return;
+            }
           }
+        }
+        else if (!conflict.srcLeaf && conflict.dstLeaf) {
+          conflictStr += ">>>>Existing key is not a file\n" + conflict.key + ": " + conflict.value + "\n";
+          if (this.conflictsResolution == "abort") abortImport = true;
+          if (this.sameKeysResolution == "overwrite") {
+            var response: string | undefined;
+            var noOwPromptCR = this.context.globalState.get("no_overwrite_prompt_for_conflicting_keys");
+            if (!noOwPromptCR) {
+              await vscode.window.showWarningMessage("Etcd Explorer: You have chosen to overwrite the values for same keys in settings, do you wish to continue?", { "modal": true }, "Yes, and don't ask again", "Yes").then((s) => { response = s; });
+              if (response === undefined) abortImport = true;
+              else if (response != "Yes") {
+                this.context.globalState.update("no_overwrite_prompt_for_conflicting_keys", true);
+              }
+            }
+          }
+        }
+        else {
+          conflictStr += ">>>>Key already exists\n" + conflict.key + ": " + conflict.value + "\n";
+          if (this.sameKeysResolution == "abort") abortImport = true;
+          if (this.sameKeysResolution == "overwrite") {
+            var response: string | undefined;
+            var noOwPromptSK = this.context.globalState.get("no_overwrite_prompt_for_same_keys");
+            if (!noOwPromptSK) {
+              await vscode.window.showWarningMessage("Etcd Explorer: You have chosen to overwrite the values for same keys in settings, do you wish to continue?", { "modal": true }, "Yes, and don't ask again", "Yes").then((s) => { response = s; });
+              if (response === undefined) abortImport = true;
+              else if (response != "Yes") {
+                this.context.globalState.update("no_overwrite_prompt_for_same_keys", true);
+              }
+            }
+          }
+        }
+        if (abortImport)
+          continue;
+        if (conflict.srcLeaf != conflict.dstLeaf) {
           if (this.conflictsResolution == "ignore")
             continue;
           if (this.conflictsResolution == "overwrite") {
+            // showWarningMessage
             if (!conflict.srcLeaf) {
-              var n = this.findEtcdNode(conflict.key.replace('/' + separator + '$/', ""));
+              var n = this.findEtcdNode(conflict.key.replace(new RegExp(separator + '*$'), ""));
               if (n) {
                 var parent = n.Parent();
                 this.deleteKeys(n.prefix).then(() => {
@@ -165,7 +220,7 @@ export class EtcdExplorerBase {
                     parent.getChildren().removeNode(n);
                     parent.refreshChildren = true;
                   }
-                });
+                }).catch(() => { });
               }
             }
             else
@@ -173,28 +228,22 @@ export class EtcdExplorerBase {
           }
         }
         else {
-          if (this.sameKeysResolution == "abort") {
-            vscode.window.showErrorMessage("Aborting json import as some keys already exist");
-            var p = new Promise(async (resolve) => {
-              var str = "Conflicts:\n ==================================\n";
-              for (var c of conflicts) {
-                str += c.key + ": " + c.value + "\n";
-              }
-              let doc = await vscode.workspace.openTextDocument({ content: str, language: "json" });
-              vscode.window.showTextDocument(doc, { preview: false });
-              resolve();
-            });
-
-            return;
-          }
           if (this.sameKeysResolution == "ignore")
             continue;
           if (this.sameKeysResolution == "overwrite")
             writes.push({ key: conflict.key, value: conflict.value });
         }
       }
+      if (abortImport) {
+        let doc = await vscode.workspace.openTextDocument({ content: conflictStr, language: "json" });
+        vscode.window.showTextDocument(doc, { preview: false });
+        vscode.window.showErrorMessage("Aborting json import due to conflict with existing keys");
+        return;
+      }
     }
+    if (token && token.isCancellationRequested) return;
     for (var write of writes) {
+      if (token && token.isCancellationRequested) return;
       var n = this.findEtcdNode(write.key);
       if (n) {
         var parent = n.Parent();
@@ -204,78 +253,120 @@ export class EtcdExplorerBase {
           parent.refreshChildren = true;
         }
       }
-      var self = this;
-      await this.deleteKeys(write.key);
-      self.write(write.key, write.value);
+      await this.deleteKeys(write.key).then(() => { }).catch((reason) => { });
+      await this.write(write.key, write.value).then(() => { }).catch((reason) => {
+        vscode.window.showErrorMessage("Error: while writing key " + write.key + " [" + reason + "]");
+      });
     }
   }
 
-  jsonToLevelNodeList(jsonObj: JSON, node: EtcdNode) {
+  protected isValueType(value: any): boolean {
+    var str = Object.prototype.toString.call(value);
+    var type = str.slice(8, -1).toLowerCase();
+    var isValType = false;
+    if (type == "string"
+      || type == "boolean"
+      || type == "number"
+      || type == "date"
+      || type == "array"
+      || type == "undefined"
+      || type == "null") {
+      isValType = true;
+    }
+    else {
+      isValType = false;
+    }
 
+    return isValType;
+  }
+
+  protected getType(value: any): string {
+    var str = Object.prototype.toString.call(value);
+    var type = str.slice(8, -1).toLowerCase();
+    return type;
+  }
+
+  async jsonToLevelNodeList(jsonObj: JSON, node: EtcdNode, explorer: EtcdExplorerBase) {
+
+    // use stack to iterate DFS 
     var stack = Array<{ json: JSON, node: EtcdNode }>();
     stack.push({ json: jsonObj, node: node });
-
+    var staleMap = new HashMap();
     while (stack.length > 0) {
       var obj = stack.pop();
       if (!obj) break;
       var entries = Object.entries(obj.json);
       var nodeList = obj.node.getChildren();
-      var currentLabels = [];
       for (var chNode of nodeList.toArray()) {
-        chNode.stale = true;
-        currentLabels.push(chNode.label);
+        if (!staleMap.has(chNode.prefix)) {
+          chNode.stale = true;
+          staleMap[chNode.prefix] = chNode;
+        }
       }
       for (var entry of entries) {
         var isLeaf = false;
         var key = entry[0];
         var value = entry[1];
-        var str = Object.prototype.toString.call(value);
-        var type = str.slice(8, -1).toLowerCase();
-        if (type == "string"
-          || type == "boolean"
-          || type == "number"
-          || type == "date"
-          || type == "array"
-          || type == "undefined"
-          || type == "null"
-        ) {
-          isLeaf = true;
-        }
-        else {
-          isLeaf = false;
-        }
-        if (currentLabels.indexOf(key) > -1) {
-          var existingNode = nodeList.getNode(key);
-          if (existingNode != undefined && existingNode.isLeafNode() == isLeaf) {
-            existingNode.stale = false;
-            continue;
+        isLeaf = explorer.isValueType(value);
+        var existingNodes = nodeList.getNodeByLabel(key);
+        var existingLeafNode = existingNodes.leaf;
+        var existingDirNode = existingNodes.dir;
+        if ((isLeaf && existingLeafNode) || (!isLeaf && existingDirNode)) {
+          if (isLeaf) {
+            existingLeafNode?.setData(value); // just update the value
+            if (existingLeafNode) existingLeafNode.stale = false; // nothing else to change here
           }
           else {
-            nodeList.removeNode(existingNode);
+            if (existingDirNode) {
+              existingDirNode.stale = false; // nothing to change here
+              stack.push({ json: value, node: existingDirNode });
+            }
+          }
+          // leave the node as is
+          continue;
+        }
+        else {
+          if (explorer.schema().startsWith('etcd2')) {
+            // remove the current node and it'll b replace by new node
+            // replacing dir by leaf and leaf by dir
+            nodeList.removeNode(isLeaf ? existingDirNode : existingLeafNode);
           }
         }
+
         if (isLeaf) {
           var prefix = obj.node.prefix + key;
-          var newNode = new EtcdLeafNode(key, prefix, this, obj.node, value);
-          nodeList.pushNode(newNode);
+          var parent: EtcdNode | undefined = obj.node;
+
+          /*          if (key == '__EtcdExplorerSpecialCaseVal__') {
+                      key = obj.node.label;
+                      prefix = obj.node.prefix.replace(new RegExp(separator + '*$'), "");
+                      parent = (obj.node.Parent()) ? obj.node.Parent() : explorer.RootNode();
+                    }
+          */
+          if (explorer.getType(value) == "array") {
+            for (var data of value) {
+              var nObj = Object.create({});
+              nObj[key] = data;
+              stack.push({ json: nObj, node: obj.node });
+            }
+            continue;
+          }
+          var newNode = new EtcdLeafNode(key, prefix, explorer, parent, value);
+          parent?.getChildren().pushNode(newNode);
         }
         else {
           var prefix = obj.node.prefix + key + separator;
-          var newNode = new EtcdNode(key, prefix, this, obj.node);
+          var newNode = new EtcdNode(key, prefix, explorer, obj.node);
           nodeList.pushNode(newNode);
           stack.push({ json: value, node: newNode });
         }
       }
 
-      var removal = []
-      for (chNode of nodeList.toArray()) {
-        if (chNode.stale)
-          removal.push(chNode);
+      // now remove nodes that are stale
+      for (var nodeKey of staleMap.keys()) {
+        if (staleMap[nodeKey].stale)
+          nodeList.removeNode(staleMap[nodeKey]);
       }
-      for (var chNode of removal) {
-        nodeList.removeNode(chNode);
-      }
-      nodeList.updatingNodes = false;
     }
   }
 
@@ -319,7 +410,7 @@ export class EtcdExplorerBase {
     if (node instanceof EtcdSpecialNode) {
       return;
     }
-    node.Value().then(async (value: any) => {
+    node.updateValue().then(async (value: any) => {
       //let doc = await vscode.workspace.openTextDocument({ content: value, language: "json" }); // calls back into the provider 
       //var token = require('crypto').randomBytes(48).toString('hex');
       //token = token.toString('base64').replace(/\//g, '_').replace(/:/g, '-');
@@ -332,25 +423,37 @@ export class EtcdExplorerBase {
     });
   }
 
-  deleteKeys(prefix: string): Thenable<void> {
-    return Promise.resolve();
+  async deleteKeys(prefix: string) {
+    return new Promise((resolve, reject) => { resolve(); });
   }
 
-  initAllData(node: EtcdNode, callback: Function, ignoreParentKeys?: boolean, recursive?: boolean) { }
+  async initAllData(node: EtcdNode, callback: Function, ignoreParentKeys?: boolean, recursive?: boolean) {
+  }
 
   async importResource() {
     var self = this;
     var promise = vscode.window.showOpenDialog({ openLabel: "Open JSON File", canSelectFiles: true, canSelectFolders: false, canSelectMany: false, filters: { "Json Files": ["json"] } });
     promise.then(
-      (jsonFile) => {
-        //console.log(jsonFile);
+      async (jsonFile) => {
+
         if (jsonFile && jsonFile.length > 0) {
-          var path = jsonFile[0].fsPath;
-          const fs = require('fs');
-          let rawdata = fs.readFileSync(path);
-          let jsonObj = JSON.parse(rawdata);
-          self.jsonToEtcd(jsonObj);
-          self.refresh();
+          vscode.window.withProgress({
+            location: vscode.ProgressLocation.Notification,
+            title: "importing JSON to etcd",
+            cancellable: true
+          }, (progress, token) => {
+            return new Promise(async (resolve) => {
+              progress.report({ message: "importing " + jsonFile[0].fsPath + " ..." });
+              //console.log(jsonFile);
+              var path = jsonFile[0].fsPath;
+              const fs = require('fs');
+              let rawdata = fs.readFileSync(path);
+              let jsonObj = JSON.parse(rawdata);
+              await self.jsonToEtcd(jsonObj, token);
+              self.refreshData();
+              resolve();
+            });
+          });
         }
       }
     );
@@ -367,7 +470,7 @@ export class EtcdExplorerBase {
       var valueBox = vscode.window.createInputBox();
       valueBox.title = "Add Key Value";
       valueBox.prompt = "Please type your value here?";
-      valueBox.onDidAccept(() => {
+      valueBox.onDidAccept(async () => {
         valueBox.hide();
         var value = valueBox.value;
         var key = keyBox.value;
@@ -385,7 +488,8 @@ export class EtcdExplorerBase {
           newObj[label] = jsonObj;
           jsonObj = newObj;
         }
-        self.jsonToEtcd(jsonObj);
+        await self.jsonToEtcd(jsonObj);
+        self.refreshData();
         //self.write(key, value);
         //console.log(inputBox.value + ": " + inputBox2.value);
       });
@@ -394,31 +498,59 @@ export class EtcdExplorerBase {
     keyBox.show();
   }
 
+  async jsonToTextDocument(jsonObj: JSON, node: EtcdNode, explorer: EtcdExplorerBase) {
+    var conf = explorer.getConfig();
+    var str = "";
+    if (conf.exportJSON.nested) {
+      str = JSON.stringify(jsonObj);
+    }
+    else {
+      str = "{";
+      if (node.isLeafNode()) {
+        await node.updateValue();
+        str += "\"" + node.prefix + "\": \"" + node.getData() + "\"";
+      }
+      else {
+        var stack = new Array<EtcdNode>();
+        stack.push(node);
+        while (stack.length > 0) {
+          var n = stack.pop();
+          for (var ch of n?.getChildren().toArray()) {
+            if (ch.isLeafNode()) {
+              await ch.updateValue();
+              str += ((str == "{") ? "\"" : ", \"") + ch.prefix + "\": \"" + ch.getData() + "\"";
+            }
+            else {
+              stack.push(ch);
+            }
+          }
+        }
+      }
+      str += "}";
+    }
+    let doc = await vscode.workspace.openTextDocument({ content: str, language: "json" });
+    vscode.window.showTextDocument(doc, { preview: false });
+  }
+
+
   async exportResource(nodeResource?: EtcdNode) {
     var node = (nodeResource != undefined) ? nodeResource : this.RootNode();
     if (node instanceof EtcdSpecialNode) {
       return;
     }
-
-    var promise = vscode.window.withProgress({
+    var self = this;
+    vscode.window.withProgress({
       location: vscode.ProgressLocation.Notification,
       title: "exporting etcd data to JSON",
       cancellable: true
     }, (progress, token) => {
-      progress.report({ message: "loading data for " + node.label + " ..." });
-      var count = 0;
-      var p = new Promise((resolve, reject) => {
-        this.initAllData(node, async (jsonObj: JSON, etcdNode: EtcdNode) => {
-          var str = JSON.stringify(jsonObj);
-          let doc = await vscode.workspace.openTextDocument({ content: str, language: "json" });
-          vscode.window.showTextDocument(doc, { preview: false });
-          resolve();
-        }, false, true);
+      return new Promise((resolve) => {
+        progress.report({ message: "loading data for " + node.label + " ..." });
+        var count = 0;
+        self.initAllData(node, this.jsonToTextDocument, false, true);
+        resolve();
       });
-      return p;
     });
-    promise.then(async () => {
-    }, (reason) => { console.log(reason); });
   }
 
   async copyResourcePrefix(nodeResource?: EtcdNode) {
@@ -432,36 +564,66 @@ export class EtcdExplorerBase {
   }
 
   async deleteResource(node: EtcdNode) {
-    if (node === undefined) return;
-    if (node instanceof EtcdSpecialNode) {
+    if (node != undefined && node instanceof EtcdSpecialNode) {
       return;
     }
-    var prompt = vscode.window.showWarningMessage("Are you sure, you wish to delete all keys with prefix " + node.prefix, "Yes", "No");
-    prompt.then((value) => {
-      if (value == "Yes") {
-        var promise = this.deleteKeys(node.prefix);
-        promise.then(() => {
-          var parent = node.Parent();
-          if (parent != undefined) {
-            parent.getChildren().removeNode(node);
-            parent.refreshChildren = true;
-            this.refresh();
-          }
-        });
-      }
-    });
+    var self = this;
+    var delKeys = function (self: any, key: string) {
+      var prompt = vscode.window.showWarningMessage("Are you sure, you wish to delete all keys with prefix " + key, "Yes", "No");
+      prompt.then((value) => {
+        if (value == "Yes") {
+          self.deleteKeys(key).then(() => {
+            if (!node) {
+              node = self.findEtcdNode(key);
+            }
+            node = node ? node : self.RootNode();
+            var parent = node.Parent();
+            if (parent != undefined) {
+              parent.getChildren().removeNode(node);
+              parent.refreshChildren = true;
+              self.refresh();
+            }
+            if (node == self.RootNode()) {
+              node.getChildren().clearChildren();
+              self.refreshView();
+            }
+          }).catch(() => {
+            vscode.window.showErrorMessage("Failed to delete keys with prefix " + key);
+          });
+        }
+      });
+    }
+    if (!node) {
+      var keyBox = vscode.window.createInputBox();
+      keyBox.title = "Delete Key";
+      keyBox.onDidAccept(() => {
+        keyBox.hide();
+        var key = keyBox.value;
+        delKeys(self, key);
+      });
+      keyBox.show();
+    }
+    else {
+      delKeys(self, node.prefix);
+    }
   }
 
   async getValue(key: string): Promise<any> { }
 
   findEtcdNode(key: string, nodeList?: EtcdNodeList | undefined): EtcdNode | undefined {
     var nodes = nodeList ? nodeList : this.RootNode().getChildren(true);
-    var labels = key.split(separator);
+    key = key.replace(new RegExp(separator + '*$'), "");
+    var labels = key.split(separator).reverse();
     var node: EtcdNode | undefined;
-    for (var label of labels) {
+    while (labels.length > 0) {
+      var label = labels.pop();
       if (!label || label.length == 0) continue;
       if (!nodes.hasLabel(label)) return undefined;
-      var node = nodes.getNode(label);
+      var nodeObj = nodes.getNodeByLabel(label);
+      node = nodeObj.dir;
+      if (labels.length == 0) {
+        node = key.endsWith(separator) ? nodeObj.dir : nodeObj.leaf;
+      }
       if (!node) return undefined;
       nodes = node.getChildren();
     }
@@ -479,7 +641,7 @@ export class EtcdNode extends vscode.TreeItem {
   protected data?: string;
   public stale = false;
   public uri: string;
-  public specialKey: string;
+
   constructor(
     public readonly label: string,
     public readonly prefix: string,
@@ -487,7 +649,7 @@ export class EtcdNode extends vscode.TreeItem {
     parentNode?: EtcdNode
   ) {
     super(label, vscode.TreeItemCollapsibleState.Collapsed);
-    this.specialKey = label;
+    this.id = prefix;
     this.isLeaf = false;
     this.explorer = parentNode ? parentNode.explorer : etcd_explorer;
     this.children = new EtcdNodeList(this.explorer);
@@ -512,7 +674,7 @@ export class EtcdNode extends vscode.TreeItem {
     return this.data;
   }
 
-  async Value(): Promise<any> {
+  async updateValue(): Promise<any> {
     var data: any;
     var error: any;
     if (!this.isLeafNode())
@@ -537,13 +699,11 @@ export class EtcdNode extends vscode.TreeItem {
     });
   }
 
-  setValue(val: string) {
+  setData(val: string) {
     this.data = val;
     let uri = vscode.Uri.parse(this.explorer.schema() + ":" + this.prefix);
     this.explorer.refreshDocument(uri);
   }
-
-
 
   getChildren(refreshCheck?: boolean): EtcdNodeList {
     var refresh = refreshCheck ? refreshCheck : false;
@@ -552,38 +712,6 @@ export class EtcdNode extends vscode.TreeItem {
       this.refreshChildren = false;
     }
     return this.children;
-  }
-
-  nodeJSONObj(): Object {
-    var nodes = this.getChildren().toArray();
-    var ready = !this.getChildren().updatingNodes;
-    const out = Object.create(null);
-
-    if (ready) {
-      if (this.isLeaf) {
-        out[this.label] = this.data;
-      }
-      else {
-        nodes = this.getChildren().toArray();
-        if (nodes.length == 0) {
-        }
-        else if (nodes.length == 1) {
-          var obj = nodes[0].nodeJSONObj();
-          out[this.label] = obj;
-        }
-        else {
-          var arr = Object.create(null)
-          for (var n of this.getChildren().toArray()) {
-            var obj = n.nodeJSONObj();
-            for (var entry of Object.entries(obj)) {
-              arr[entry[0]] = entry[1];
-            }
-          }
-          out[this.label] = arr;
-        }
-      }
-    }
-    return out;
   }
 }
 
@@ -606,7 +734,7 @@ export class EtcdLeafNode extends EtcdNode {
 export class EtcdRootNode extends EtcdNode {
   constructor(etcd_explorer: EtcdExplorerBase) {
     super("Root", separator, etcd_explorer);
-    this.specialKey = separator + this.label;
+    this.id = "root" + separator + this.label;
   }
   contextValue = 'etcdrootnode';
 }
@@ -618,7 +746,11 @@ export class EtcdSpecialNode extends EtcdLeafNode {
       etcd_explorer,
       parent
     );
-    this.specialKey = separator + this.label;
+    this.isLeaf = true;
+    // add random token to make id unique
+    var token = require('crypto').randomBytes(48).toString('hex');
+    token = token.toString('base64').replace(/\//g, '_').replace(/:/g, '-');
+    this.id = "special" + separator + this.label + separator + token;
   }
   contextValue = 'specialetcdnode';
 }
@@ -647,9 +779,10 @@ export class EtcdUpdatingNode extends EtcdSpecialNode {
 
 export class EtcdNodeList {
   protected nodeMap = new HashMap();
-  public updatingNodes: boolean;
+  protected updatingNodes: boolean;
   public pageCount = 1;
   private explorer: EtcdExplorerBase;
+  private lastUpdated = Date.now() - 10000;
   constructor(etcd_explorer: EtcdExplorerBase) {
     this.updatingNodes = false;
     this.explorer = etcd_explorer;
@@ -662,68 +795,127 @@ export class EtcdNodeList {
     this.nodeMap.clear();
   }
 
+  isUpdating() {
+    return this.updatingNodes;
+  }
+
+  canUpdate() {
+    return (this.updatingNodes == false) && ((Date.now() - this.lastUpdated) > 500);
+  }
+
+  updating() {
+    this.updatingNodes = true;
+  }
+
+  updated() {
+    this.updatingNodes = false;
+    this.lastUpdated = Date.now();
+  }
+
   removeNode(node?: EtcdNode) {
     if (node != undefined) {
       var p = node.Parent();
       if (p)
         p.refreshChildren = true;
-      this.nodeMap.delete(node.specialKey);
+      var key = node.isLeafNode() ? "leaf" + separator + node.label : "dir" + separator + node.label;
+      if (node.id?.startsWith('special')) key = node.id;
+      this.nodeMap.delete(key);
     }
   }
 
   removeUpdatingNodes() {
-    this.nodeMap.delete(separator + "loading");
+    var spKeys = new Array<string>();
+    this.nodeMap.forEach(function (value: any, key: string) {
+      if (key.startsWith("special" + separator + "loading"))
+        spKeys.push(key);
+    });
+    for (var key of spKeys) {
+      this.nodeMap.delete(key);
+    }
   }
 
   removeSpecialNodes() {
-    this.nodeMap.delete(separator + "Special");
-    this.nodeMap.delete(separator + "next->");
-    this.nodeMap.delete(separator + "Empty Workspace");
-    this.nodeMap.delete(separator + "loading");
+    var spKeys = new Array<string>();
+    this.nodeMap.forEach(function (value: any, key: string) {
+      if (key.startsWith("special" + separator))
+        spKeys.push(key);
+    });
+    for (var key of spKeys) {
+      this.nodeMap.delete(key);
+    }
   }
 
   countSpecialNodes() {
     var spNodes = 0;
-    var sp = [
-      separator + "Special",
-      separator + "next->",
-      separator + "Empty Workspace",
-      separator + "loading"
-    ];
-    for (var spLabel of sp) {
-      if (this.nodeMap.has(spLabel)) {
+    this.nodeMap.forEach(function (value: any, key: string) {
+      if (key.startsWith("special" + separator))
         spNodes++;
-      }
-    }
+    });
     return spNodes;
   }
 
-  getNode(label: string): EtcdNode | undefined {
-    var result = undefined;
-    result = this.nodeMap.get(label);
+  getNodeByLabel(label: string): { dir: EtcdNode | undefined, leaf: EtcdNode | undefined } {
+    var result = Object.create({});
+    result["dir"] = this.nodeMap.get("dir" + separator + label);
+    result["leaf"] = this.nodeMap.get("leaf" + separator + label);
+    return result;
+  }
+
+  getNodeById(nodeId: string): EtcdNode | undefined {
+    var result: EtcdNode | undefined;
+    var labels = nodeId.split(separator);
+    while (labels.length > 0) {
+      var label = labels.pop();
+      if (label && label.length > 0) {
+        var val = this.getNodeByLabel(label);
+        result = nodeId.endsWith(separator) ? val.dir : val.leaf;
+        break;
+      }
+    }
     return result;
   }
 
   pushNode(node: EtcdNode) {
-    this.nodeMap.set(node.specialKey, node);
+    var key = node.isLeafNode() ? "leaf" + separator + node.label : "dir" + separator + node.label;
+    if (node.id?.startsWith('special')) key = node.id;
+    this.nodeMap.set(key, node);
   }
 
-  pushLabel(label: string, pre: string, node: EtcdNode, isleaf?: boolean, value?: string) {
+  pushLabel(label: string, prefix: string, node: EtcdNode, isleaf?: boolean, value?: string) {
+    var key = node.isLeafNode() ? "leaf" + separator + node.label : "dir" + separator + node.label;
+    if (node.id?.startsWith('special')) key = node.id;
     if (isleaf) {
-      this.nodeMap.set(label, new EtcdLeafNode(label, pre, this.explorer, node, value));
+      this.nodeMap.set(key, new EtcdLeafNode(label, prefix, this.explorer, node, value));
     }
     else {
-      this.nodeMap.set(label, new EtcdNode(label, pre, this.explorer, node));
+      this.nodeMap.set(key, new EtcdNode(label, prefix, this.explorer, node));
     }
   }
 
 
   hasNode(node: EtcdNode): boolean {
-    return this.nodeMap.has(node.specialKey);
+    var key = node.isLeafNode() ? "leaf" + separator + node.label : "dir" + separator + node.label;
+    if (node.id?.startsWith('special')) key = node.id;
+    return this.nodeMap.has(key);
   }
 
   hasLabel(label: string): boolean {
-    return this.nodeMap.has(label);
+    return this.nodeMap.has("leaf" + separator + label) || this.nodeMap.has("dir" + separator + label);
+  }
+
+  hasId(nodeId: string): boolean {
+    var key = nodeId
+    if (!nodeId.startsWith('special')) {
+      var labels = nodeId.split(separator);
+      while (labels.length > 0) {
+        var label = labels.pop();
+        if (label && label.length > 0) {
+          var key = nodeId.endsWith(separator) ? "dir" + separator + label : "leaf" + separator + label;
+          break;
+        }
+      }
+    }
+    return this.nodeMap.has(key);
   }
 
   map() {